add builtins to compilation
